esycord 1.6.6
Created a new `CLI` for modification of `Data` class
generated databases with sqlite3. Access this by
running the library with the `--data` argument.

Fixed `Bot.run` method logic.

Seperated `Data` class into a new file due to size
issues.

Minor modifications to the code.

Added a new argument to run the library `-nuc` so
the library dosen't check for updates when ran in
a terminal.